+++ wakamiti-plugins/wakamiti-groovy/src/test/java/es/iti/wakamiti/groovy/GroovyStepContributorTest.java
@@ -5,16 +5,35 @@
  */
 package es.iti.wakamiti.groovy;
 
+import es.iti.wakamiti.api.Backend;
 import es.iti.wakamiti.api.WakamitiException;
+import es.iti.wakamiti.api.WakamitiStepRunContext;
 import es.iti.wakamiti.api.plan.Document;
+import org.junit.Before;
 import org.junit.Test;
 
+import java.util.LinkedHashMap;
+import java.util.LinkedList;
+import java.util.Map;
+
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 
 public class GroovyStepContributorTest {
 
     private final GroovyStepContributor contributor = new GroovyStepContributor();
+
+    @Before
+    public void setup() {
+        Map<String, Object> properties = new LinkedHashMap<>();
+        properties.put("results", new LinkedList<>());
+        properties.put("id", "ID-01");
+        Backend backend = mock(Backend.class);
+        when(backend.getExtraProperties()).thenReturn(properties);
+        WakamitiStepRunContext.set(new WakamitiStepRunContext(null, backend, null, null));
+    }
 
     @Test
     public void testExecute() {