+++ wakamiti-plugins/wakamiti-groovy/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][1],
 and this project adheres to [Semantic Versioning][2].
 
+## [Unreleased]
+
+### Added
+- Extra properties in context object
+
+
 ## [1.0.0] - 2022-12-10
 
 Initial release.