+++ wakamiti/wakamiti-api/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning][2].
 ### Added
 
 - Gpath expressions in json and xml evaluators
+- Extra properties in Backend context
 
 
 ## [1.5.1] - 2023-04-20