+++ wakamiti/wakamiti-core/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][1],
 and this project adheres to [Semantic Versioning][2].
 
+## [Unreleased]
+
+### Added
+- Extra properties in Backend context
+
+
 ## [1.7.0] - 2023-05-03
 
 ### Added